scss refactor
