+++ package.json
@@ -21,7 +21,8 @@
     "prisma": "^5.18.0",
     "react": "^18",
     "react-dom": "^18",
-    "react-google-fonts": "^1.2.1"
+    "react-google-fonts": "^1.2.1",
+    "sass": "^1.80.6"
   },
   "devDependencies": {
     "@types/node": "20.12.7",
